// IM_FRONTENDA/src/app/models/VistaProveedor.ts
export class VistaProveedor {
    proveedorId: number;
    nombre: string | null;
    tipoDocumento: number | null;
    tipoDocumento1: string | null;
    numeroDocumento: string | null;
    direccion: string | null;
    telefono: string | null;
    email: string | null;

constructor(){

    this.proveedorId=0;
    this.nombre="";
    this.tipoDocumento=0;
    this.tipoDocumento1="";
    this.numeroDocumento="";
    this.direccion="";
    this.telefono="";
    this.email="";
}

}